requirements.txt
alabaster==0.7.16
asgiref==3.7.2
Babel==2.14.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
djangorestframework==3.14.0
docutils==0.20.1
idna==3.6
imagesize==1.4.1
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
Pygments==2.17.2
pytz==2024.1
requests==2.31.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.4.4
urllib3==2.2.1
